+++ resources/rehash.rb
@@ -21,6 +21,7 @@
 # limitations under the License.
 #
 provides :rbenv_rehash
+unified_mode true
 
 property :user,       String
 property :root_path,  String, default: lazy { Chef::Rbenv::Helpers.root_path(node, user) }